\name{print.nom.uncertainty}
\alias{print.nom.uncertainty}
\title{Print method for class nom.uncertainty}
\usage{
  \method{print}{nom.uncertainty} (x, digits = 3, ...)
}
\arguments{
  \item{x}{Object to be printed.}

  \item{digits}{Number of decimal places to round to.}

  \item{...}{Not evaluated.}
}
\description{
  Print method for class nom.uncertainty
}
\keyword{internal}

