% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ryouready-package.r
\name{ryouready}
\alias{ryouready}
\title{\pkg{ryouready}: An R package to accompany the 'R you ready?' R book}
\description{
This package contains several functions that make an R beginners
life easier and which are used in our forthcoming 'R you ready?' R book.
}
\details{
The book's website where you can find the fulltext freely available
currently is \url{http://ryouready.markheckmann.de}.
}
\author{
Mark Heckmann.
}
\keyword{package}
\keyword{ryouready}

