% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{d.ngo}
\alias{d.ngo}
\title{NGO Dataset}
\description{
Data set used by Kähler (2008).
}
\references{
Kähler, W.-M. (2008). \emph{Statistische Datenanalyse: Verfahren verstehen
und mit SPSS gekonnt einsetzen}. Wiesbaden: Vieweg.
}
\keyword{data}

