\name{foo}
\alias{foo}
\title{A function to demonstrate how an R function is defined.}
\usage{
  foo(x, y)
}
\arguments{
  \item{x}{Numeric.}

  \item{y}{Numeric.}
}
\value{
  The sum of \code{x} and \code{y}.
}
\description{
  A function to demonstrate how an R function is defined.
}
\author{
  Mark Heckmann
}

