% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{foo}
\alias{foo}
\title{A function to demonstrate how an R function is defined.}
\usage{
foo(x, y)
}
\arguments{
\item{x}{Numeric.}

\item{y}{Numeric.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
A function to demonstrate how an R function is defined.
}
\author{
Mark Heckmann
}

