% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qq-plot.R
\name{ggplot.qqnorm.spss}
\alias{ggplot.qqnorm.spss}
\title{Plot the output from \code{qqplot.spss} using \code{ggplot2}}
\usage{
\method{ggplot}{qqnorm.spss}(x, plottype = 1, line = TRUE,
  l.col = "black", ...)
}
\arguments{
\item{x}{An object as returned by \code{\link{qqnorm_spss}}}

\item{plottype}{The type of plot created.
\code{1 =} Standard QQ-plot, \code{2 =} Detrended QQ-plot.}

\item{line}{Whether to plot a QQ-line (defaul is \code{TRUE})}

\item{l.col}{Color of the QQ-line.}

\item{...}{Not evaluated.}
}
\value{
A ggplot object.
}
\description{
Plot the output from \code{qqplot.spss} using \code{ggplot2}
}
\keyword{internal}

