\name{print.ord.gamma}
\alias{print.ord.gamma}
\title{Print method for class ord.gamma}
\usage{
  \method{print}{ord.gamma} (x, digits = 3, ...)
}
\arguments{
  \item{x}{Object to be printed.}

  \item{digits}{Number of decimal places to round to.}

  \item{...}{Not evaluated.}
}
\description{
  Print method for class ord.gamma
}
\keyword{internal}

