\name{print.ord.somersd}
\alias{print.ord.somersd}
\title{Print method for class somersd}
\usage{
  \method{print}{ord.somersd} (x, digits = 3, ...)
}
\arguments{
  \item{x}{Object to be printed.}

  \item{digits}{Number of decimal places to round to.}

  \item{...}{Not evaluated.}
}
\description{
  Print method for class somersd
}
\keyword{internal}

