\docType{package}
\name{ryouready}
\alias{ryouready}
\alias{ryouready-package}
\title{\pkg{ryouready}: An R package to accompany the 'R you ready?' R book}
\description{
  This package contains several functions that make an R
  beginners life easier and which are used in my
  forthcoming 'R you ready?' R book.
}
\details{
  The book's website where you can find the fulltext freely
  available currently is
  \url{http://ryouready.markheckmann.de}.
}
\author{
  Mark Heckmann.
}
\keyword{internal}
\keyword{package}
\keyword{ryouready}

